+++ RELEASE_NOTES.md
@@ -4,7 +4,7 @@
 * commons-lib upgrade
 * tomcat scripts updates
 * GMB-1062 - Provision to enter the socket timeout in http sync processor - updated socket timeout validation
-* GMB-1073 - Mailbox created successfully with invalid tenancy key through postman services - reverted
+* GMB-1073 - Mailbox created successfully with invalid tenancy key through postman services
 * GMB-1064, GMB-1065 - JSCH 0.1.54 causing issues for the kex alg diffie-hellman-group-exchange-sha1 
 
 5.0.1